Improve ArrayTypeHandler
ArrayTypeHandler unwraps the contents of the java.sql.Array when fetching data. Upon insertion it expects an instance of java.sql.Array as the parameter.

This changeset improves ArrayTypeHandler so that it accepts an Object[] parameter for insert, constructing the java.sql.Array object if necessary. This aproach relies on the jdbcType of the column - it must reflect the array element's type. For setting NULL values on ARRAY columns I had to refactor BaseTypeHandler a bit - but I'm quite sure that this won't break anything.